add get_bsd_hostname() and has_domain()
get_bsd_hostname takes a string and returns a proper FQDN to be
written to /etc/myname or NULL if the passed string is unsalvagable.
if passed string does not contain a domain, the current domain from
/etc/myname is used. if /etc/myname does not contain a trailing
domain, an error message is printed and the default (".home.network")
is used.